fix(grid): move months method from Activity to Grid #48
